backend/requirements.txt
streamlit
pandas
numpy
pdfplumber
statsmodels
scikit-learn
joblib
cryptography
web3
ibm-watsonx-ai
matplotlib
python-dotenv
reportlab
